+++ graphql-maven-plugin-logic/src/main/java/com/graphql_java_generator/plugin/generate_schema/GenerateGraphQLSchema.java
@@ -156,7 +156,7 @@ public class GenerateGraphQLSchema {
 		if (configuration.getTemplates().containsKey(template.name())) {
 			return configuration.getTemplates().get(template.name());
 		} else {
-			return template.getDefaultValue();
+			return template.getDefaultPath();
 		}
 	}